pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiogram>=3.22.0",
    "aiosqlite>=0.21.0",
    "matplotlib>=3.10.6",
    "openai>=1.108.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
